requirements.txt
Babel==1.3
PyYAML==3.11
amqp==1.4.6
anyjson==0.3.3
argparse==1.2.1
eventlet==0.15.2
greenlet==0.4.6
iso8601==0.1.10
kombu==3.0.23
msgpack-python==0.4.6
netaddr==0.7.12
netifaces==0.10.4
oslo.config==1.11.0
oslo.context==0.3.0
oslo.i18n==1.6.0
oslo.messaging==1.4.1
oslo.serialization==1.5.0
oslo.utils==1.5.0
pbr==0.11.0
pytz==2014.7
six==1.9.0
stevedore==1.4.0
wsgiref==0.1.2
